Complete

+++ src/main/java/lesson1/Pool.java
@@ -11,7 +11,7 @@ public class Pool {
         return length;
     }
 
-    public void setLength(int length) {
+    public void setLength(double length) {
         this.length = length;
     }
